Restored utils.py file
